+++ Samples/UpdateLatestOSReleaseCompliancePolicy.ps1
@@ -1,11 +1,31 @@
+<#
+.SYNOPSIS
+    This script will update the minimum OS version of a compliance policy and app protection policy to the latest iOS/Android release.
 
-#set the graph endpoint
-$Global:GraphEndpoint = 'https://graph.microsoft.com'
-$PolicyId = "b79cb75a-2dd7-496e-b9af-13f0e9e2bba0"
+.DESCRIPTION
+    This script will update the minimum OS version of a compliance policy to the latest iOS/Android release.
+
+.EXAMPLE
+    .\UpdateLatestOSReleaseCompliancePolicy.ps1
+
+.NOTES
+    File Name      : UpdateLatestOSReleaseCompliancePolicy.ps1
+    Author         : Powershellcrack
+
+#>
+#SET THE POLICY IDS
+$iOSCompliancePolicyId = "712599b4-c442-49d4-867c-a92fc2da371a"
+$iOSAppProtectPolicyId = "T_2824c38c-9fa8-4f7e-af7c-e98fe2cc4c78"
+
+$AndriodCompliancePolicyId = "58aaee1b-9930-4c56-8135-722e7efa327f"
+$AndriodAppProtectPolicyId = "T_6ffb82b6-fbbd-4b68-bab8-15fa2860ca3b"
 
 #IMPORT MODULES
 #==============================================
+#Set Azure Commercial Endpoint
+$Global:GraphEndpoint = 'https://graph.microsoft.com'
 
+#install required modules
 $modules = @(
     'Microsoft.Graph.Authentication',
     'Microsoft.Graph.Applications',
@@ -24,6 +44,7 @@ foreach ($module in $modules){
         Write-Host "Done" -ForegroundColor Green
     }
 }
+
 #MAIN
 #==============================================
 
@@ -32,8 +53,21 @@ Write-Host '----------------------------------------' -ForegroundColor White
 #connect to the graph
 Write-Host "Connecting to Microsoft Graph..." -ForegroundColor White -NoNewline
 Connect-MgGraph -NoWelcome
+
+#connect to the grpah using a service principal
 #Connect-IDMGraphApp -TenantId $TenantId -ClientId $ClientId -ClientSecret $ClientSecret
 
-#Set the latest Windows 11 OS release as the minimum version for the compliance policy
-$MiniOS = Get-IDMCompliancePolicyOSRelease -Platform "Windows11" -Latest
-Update-IDMCompliancePoliciesOSVersion -OSVersionType MinimumVersion -PolicyId $PolicyId -OSVersion $MiniOS
+
+#get the latest iOS release
+$iOSLatestOS = Get-IDMCompliancePolicyOSRelease -Platform iOS -Latest
+
+#Set the latest iOS release as the minimum version for the compliance policy and app protection policy
+Update-IDMCompliancePolicyOSVersion -PolicyId $iOSCompliancePolicyId -OSVersionType MinimumVersion -OSVersion $iOSLatestOS
+Update-IDMAppProtectionPolicyOSCondition -Platform iOS -PolicyId $iOSAppProtectPolicyId -OSCondition MinimumVersion -OSVersion $iOSLatestOS
+
+#get the latest Android release
+$AndroidLatestOS = Get-IDMCompliancePolicyOSRelease -Platform Android -Latest
+
+#Set the latest Android release as the minimum version for the compliance policy and app protection policy
+Update-IDMCompliancePolicyOSVersion -PolicyId $AndriodCompliancePolicyId -OSVersionType MinimumVersion -OSVersion $AndroidLatestOS
+Update-IDMAppProtectionPolicyOSCondition -Platform Android -PolicyId $AndriodAppProtectPolicyId -OSCondition MinimumVersion -OSVersion $AndroidLatestOS